+++ pyobjc-framework-AVFoundation/PyObjCTest/test_avmetadataobject.py
@@ -41,3 +41,16 @@ class TestAVMetadataObject(TestCase):
             self.assertIsInstance(AVFoundation.AVMetadataObjectTypeDogBody, str)
         with self.subTest("salientobject"):
             self.assertIsInstance(AVFoundation.AVMetadataObjectTypeSalientObject, str)
+
+    @min_os_level("12.3")
+    def testConstants12_3(self):
+        self.assertIsInstance(AVFoundation.AVMetadataObjectTypeCodabarCode, str)
+        self.assertIsInstance(AVFoundation.AVMetadataObjectTypeGS1DataBarCode, str)
+        self.assertIsInstance(
+            AVFoundation.AVMetadataObjectTypeGS1DataBarExpandedCode, str
+        )
+        self.assertIsInstance(
+            AVFoundation.AVMetadataObjectTypeGS1DataBarLimitedCode, str
+        )
+        self.assertIsInstance(AVFoundation.AVMetadataObjectTypeMicroQRCode, str)
+        self.assertIsInstance(AVFoundation.AVMetadataObjectTypeMicroPDF417Code, str)