+++ pyobjc-framework-PassKit/PyObjCTest/test_pkconstants.py
@@ -83,3 +83,12 @@ class TestPKConstants(TestCase):
     def test_constants12_0(self):
         self.assertIsInstance(PassKit.PKPaymentNetworkWaon, str)
         self.assertIsInstance(PassKit.PKPaymentNetworkNanaco, str)
+
+    @min_os_level("12.2")
+    def test_constants12_1(self):
+        # SDK Says available in 10.1, but not available there
+        self.assertIsInstance(PassKit.PKPaymentNetworkDankort, str)
+
+    @min_os_level("12.4")
+    def test_constants12_4(self):
+        self.assertIsInstance(PassKit.PKPaymentNetworkAppleStoredValue, str)